Prevent failure when removing config file if it does not exist

+++ tests/autopilot/camera_app/tests/test_capture.py
@@ -28,7 +28,8 @@ class TestCapture(CameraAppTestCase):
     def setUp(self):
         # Remove configuration file where knowledge of the photo roll hint's necessity is stored
         config_file = os.path.expanduser("~/.config/com.ubuntu.camera/com.ubuntu.camera.conf")
-        os.remove(config_file)
+        if os.path.exists(config_file):
+            os.remove(config_file)
 
         super(TestCapture, self).setUp()
